# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        properties:
          email:
            example: user@example.com
            type: string
          first_name:
            example: John
            type: string
          id:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
          last_name:
            example: Doe
            type: string
          mobile_number:
            example: "+1234567890"
            type: string
          username:
            example: johndoe
            type: string
        type: object
    type: object
...
  handler.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: strongpassword123
        type: string
    type: object
...
  handler.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      first_name:
        example: John
        type: string
      last_name:
        example: Doe
        type: string
      mobile_number:
        example: "+1234567890"
        type: string
      password:
        example: strongpassword123
        type: string
      username:
        example: johndoe
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description:
    A comprehensive debt tracking application API with authentication,
    debt management, and multi-channel notifications.
  title: Debt Tracker API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
        - application/json
      description: Authenticate a user with email and password
      parameters:
        - description: Login credentials
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/handler.LoginRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/handler.AuthResponse"
        "400":
          description: Invalid request body
          schema:
            type: string
        "401":
          description: Invalid credentials
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Login user
      tags:
        - auth
  /auth/register:
    post:
      consumes:
        - application/json
      description: Register a new user with email and password
      parameters:
        - description: Registration details
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/handler.RegisterRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/handler.AuthResponse"
        "400":
          description: Invalid request body
          schema:
            type: string
        "409":
          description: User already exists
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Register a new user
      tags:
        - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"